requirements.txt
aiodns==4.0.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-types==0.7.0
attrs==25.4.0
beautifulsoup4==4.14.3
blinker==1.9.0
ccxt==4.5.42
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.5
click==8.3.1
colorama==0.4.6
cryptography==46.0.5
curl_cffi==0.13.0
Flask==3.0.2
frozendict==2.4.7
frozenlist==1.8.0
google-ai-generativelanguage==0.6.15
google-api-core==2.25.2
google-api-python-client==2.192.0
google-auth==2.49.0
google-auth-httplib2==0.3.0
google-generativeai==0.8.6
googleapis-common-protos==1.73.0
greenlet==3.3.2
grpcio==1.78.0
grpcio-status==1.71.2
httplib2==0.31.2
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
multidict==6.7.1
multitasking==0.0.12
numpy==2.4.2
pandas==3.0.1
peewee==4.0.1
platformdirs==4.9.4
propcache==0.4.1
proto-plus==1.27.1
protobuf==5.29.6
pyaes==1.6.1
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycares==5.0.1
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2026.1.post1
requests==2.31.0
rsa==4.9.1
setuptools==82.0.0
six==1.17.0
soupsieve==2.8.3
SQLAlchemy==2.0.48
Telethon==1.42.0
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
uritemplate==4.2.0
urllib3==2.6.3
websockets==16.0
Werkzeug==3.1.6
gunicorn==21.2.0
yarl==1.23.0
yfinance==1.2.0
